Minor localization pokes: Change three dots in FreeCell Swift's MainMenu to ellipses. Replace lostText's number specifier with the unsigned long (%lu) type.

+++ GameController.m
@@ -390,7 +390,7 @@
 
     NSBeginAlertSheet(title, defaultButton, alternateButton,
                       NSLocalizedString(@"newGameButton", @"New game button"),
-                      window, self, selector, NULL, NULL, message, moves);
+                      window, self, selector, NULL, NULL, message, (unsigned long)moves);
 }
 
 @end
